Remove analysis code dependency on scripts.distances.common
