+++ src/evaluating_rewards/analysis/distances/plot_heatmap.py
@@ -28,15 +28,14 @@ import sacred
 from evaluating_rewards import serialize
 from evaluating_rewards.analysis import stylesheets, visualize
 from evaluating_rewards.analysis.distances import aggregated, heatmaps, reward_masks
-from evaluating_rewards.distances import transitions_datasets
+from evaluating_rewards.distances import common_config
 from evaluating_rewards.scripts import script_utils
-from evaluating_rewards.scripts.distances import common
 
 plot_heatmap_ex = sacred.Experiment("plot_heatmap")
 logger = logging.getLogger("evaluating_rewards.analysis.plot_heatmap")
 
 
-transitions_datasets.make_config(plot_heatmap_ex)
+common_config.make_transitions_configs(plot_heatmap_ex)
 
 # Default configs (always applied.)
 
@@ -140,7 +139,7 @@ def _hopper_activity(args: Iterable[str]) -> bool:
 @plot_heatmap_ex.named_config
 def point_mass():
     """Heatmaps for evaluating_rewards/PointMass* environments."""
-    locals().update(**common.COMMON_CONFIGS["point_mass"])
+    locals().update(**common_config.COMMON_CONFIGS["point_mass"])
     heatmap_kwargs = {}
     heatmap_kwargs["masks"] = {
         "diagonal": [reward_masks.zero, reward_masks.same],
@@ -156,20 +155,20 @@ def point_mass():
 @plot_heatmap_ex.named_config
 def point_maze():
     """Heatmaps for imitation/PointMaze{Left,Right}-v0 environments."""
-    locals().update(**common.COMMON_CONFIGS["point_maze"])
+    locals().update(**common_config.COMMON_CONFIGS["point_maze"])
 
 
 @plot_heatmap_ex.named_config
 def point_maze_learned():
     """Compare rewards learned in PointMaze to the ground-truth reward."""
     # Analyzes models generated by `runners/transfer_point_maze.sh`.
-    locals().update(**common.COMMON_CONFIGS["point_maze_learned"])
+    locals().update(**common_config.COMMON_CONFIGS["point_maze_learned"])
 
 
 @plot_heatmap_ex.named_config
 def half_cheetah():
     """Heatmaps for seals/HalfCheetah-v0."""
-    locals().update(**common.COMMON_CONFIGS["half_cheetah"])
+    locals().update(**common_config.COMMON_CONFIGS["half_cheetah"])
     heatmap_kwargs = {
         "masks": {
             "diagonal": [reward_masks.zero, reward_masks.same],
@@ -187,7 +186,7 @@ def half_cheetah():
 @plot_heatmap_ex.named_config
 def hopper():
     """Heatmaps for seals/Hopper-v0."""
-    locals().update(**common.COMMON_CONFIGS["hopper"])
+    locals().update(**common_config.COMMON_CONFIGS["hopper"])
     heatmap_kwargs = {}
     heatmap_kwargs["masks"] = {
         "diagonal": [reward_masks.zero, reward_masks.same],
@@ -250,8 +249,8 @@ def multi_heatmaps(dissimilarities: Mapping[str, pd.Series], **kwargs) -> Mappin
 def plot_heatmap(
     vals_path: str,
     data_root: str,
-    x_reward_cfgs: Iterable[common.RewardCfg],
-    y_reward_cfgs: Iterable[common.RewardCfg],
+    x_reward_cfgs: Iterable[common_config.RewardCfg],
+    y_reward_cfgs: Iterable[common_config.RewardCfg],
     styles: Iterable[str],
     styles_for_env: Iterable[str],
     log_dir: str,
@@ -275,8 +274,8 @@ def plot_heatmap(
         save_kwargs: passed through to `analysis.save_figs`.
     """
     # Sacred turns our tuples into lists :(, undo
-    x_reward_cfgs = [common.canonicalize_reward_cfg(cfg, data_root) for cfg in x_reward_cfgs]
-    y_reward_cfgs = [common.canonicalize_reward_cfg(cfg, data_root) for cfg in y_reward_cfgs]
+    x_reward_cfgs = [common_config.canonicalize_reward_cfg(cfg, data_root) for cfg in x_reward_cfgs]
+    y_reward_cfgs = [common_config.canonicalize_reward_cfg(cfg, data_root) for cfg in y_reward_cfgs]
 
     # TODO(adam): how to specify vals_path?
     # Take `vals_path` relative to data_root, if `vals_path` not absolute
